alle testen buiten invalid delete

+++ src/main/java/be/ucll/ip/minor/groep5610/team/domain/TeamService.java
@@ -56,7 +56,10 @@ public class TeamService {
     }
 
     public void deleteTeamById(Long id) {
-        teamRepository.delete(getTeam(id));
+        System.out.println(getTeam(id));
+        if(getTeam(id) != null) {
+            teamRepository.delete(getTeam(id));
+        }
     }
 
     public List<Team> getTeamsByCategory(String category) {
